requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
faiss-cpu
langchain_chroma
langchain-groq
beautifulsoup4
sentence-transformers
langchain-huggingface
huggingface_hub
langchain_core
fastapi
uvicorn
langserve
duckdb
streamlit
ddgs